feat: 컨트랙트 충전 기능 구현
refactor: 컨트랙트 Wrapper 클래스 교체

+++ src/main/java/com/baekho/bridgenet/domain/bridge/dto/AddContractBalanceRequestDTO.java
@@ -3,10 +3,10 @@ package com.baekho.bridgenet.domain.bridge.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Getter
 public class AddContractBalanceRequestDTO {
     @NotNull
-    private BigInteger balance;
+    private BigDecimal balance;
 }
